First prototype for multiple databases

+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/pdom/PDOM.java
@@ -564,7 +564,7 @@ public class PDOM extends PlatformObject implements IPDOM {
 		fEvent.setCleared();
 	}
 
-	void reloadFromFile(File file) throws CoreException {
+	void reloadFromFile(File file, boolean deleteOld) throws CoreException {
 		assert lockCount < 0;	// must have write lock.
 		File oldFile= fPath;
 		clearCaches();
@@ -575,7 +575,9 @@ public class PDOM extends PlatformObject implements IPDOM {
 		}
 		loadDatabase(file, db.getChunkCache());
 		db.setExclusiveLock();
-		oldFile.delete();
+		if (deleteOld) {
+			oldFile.delete();	
+		}
 		fEvent.fReloaded= true;
 	}
